src/bot/pyproject.toml
[tool.poetry]
name = "telexkcdbot-bot"
version = "0.1.0"
description = ""
authors = ["coryphoenixxx <coryphoenixxx@gmail.com>"]
repository = "https://github.com/coryphoenixxx/telexkcdbot/src/bot"

[tool.poetry.dependencies]
python =  "3.10.8"
aiohttp = "^3.8.1"
aiosignal = "^1.2.0"
python-decouple = "^3.6"
aiofiles = "^22.1.0"
aiogram = "^2.22.1"
aioschedule = "^0.5.2"
beautifulsoup4 = "^4.11.1"
googletrans = "3.1.0a0"
loguru = "^0.6.0"
lxml = "^4.9.1"
numpy = "^1.23.3"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
